# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tea Backend Challenge API",
    "version": "1.0.0",
    "description": "Post feed API with relevance scoring and Redis caching",
    "contact": {
      "name": "API Support",
      "email": "support@tea-backend.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "User ID for authentication (simplified auth for demo)"
      }
    },
    "schemas": {
      "Post": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Post ID",
            "example": "64f5a1b2c3d4e5f6a7b8c9d0"
          },
          "title": {
            "type": "string",
            "description": "Post title",
            "example": "Amazing Tea Discovery"
          },
          "content": {
            "type": "string",
            "description": "Post content",
            "example": "I discovered this amazing tea blend..."
          },
          "authorId": {
            "type": "string",
            "description": "Author user ID",
            "example": "64f5a1b2c3d4e5f6a7b8c9d1"
          },
          "categoryId": {
            "$ref": "#/components/schemas/Category"
          },
          "likes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of user IDs who liked this post"
          },
          "likesCount": {
            "type": "number",
            "description": "Number of likes",
            "example": 15
          },
          "score": {
            "type": "number",
            "description": "Calculated relevance score",
            "example": 8.5
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
...
      "CreatePostRequest": {
        "type": "object",
        "required": [
          "title",
          "content",
          "categoryId"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Post title",
            "example": "Amazing Tea Discovery"
          },
          "content": {
            "type": "string",
            "description": "Post content",
            "example": "I discovered this amazing tea blend..."
          },
          "categoryId": {
            "type": "string",
            "description": "Category ID",
            "example": "64f5a1b2c3d4e5f6a7b8c9d2"
          }
        }
      },
      "BulkCreatePostsRequest": {
        "type": "object",
        "required": [
          "posts"
        ],
        "properties": {
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePostRequest"
            },
            "description": "Array of posts to create (max 50)",
            "maxItems": 50
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Request success status"
          },
          "data": {
            "type": "object",
            "description": "Response data"
          },
          "message": {
            "type": "string",
            "description": "Success message"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "PaginationResponse": {
        "type": "object",
        "properties": {
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number",
                "example": 1
              },
              "limit": {
                "type": "number",
                "example": 20
              },
              "offset": {
                "type": "number",
                "example": 0
              },
              "total": {
                "type": "number",
                "example": 100
              },
              "totalPages": {
                "type": "number",
                "example": 5
              },
              "hasNext": {
                "type": "boolean",
                "example": true
              },
              "hasPrevious": {
                "type": "boolean",
                "example": false
              },
              "nextOffset": {
                "type": "number",
                "nullable": true,
                "example": 20
              },
              "prevOffset": {
                "type": "number",
                "nullable": true,
                "example": null
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/posts": {
      "post": {
        "summary": "Create a new post",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequest"
              },
              "examples": {
                "sampleCreatePost": {
                  "summary": "A single post create example",
                  "value": {
                    "title": "Amazing Tea Discovery",
                    "content": "I discovered this amazing tea blend with notes of jasmine and honey.",
                    "categoryId": "64f5a1b2c3d4e5f6a7b8c9d2"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Post"
                        }
                      }
                    }
                  ]
                },
                "examples": {
                  "createdPost": {
                    "summary": "Example created post response",
                    "value": {
                      "success": true,
                      "data": {
                        "_id": "64f5a1b2c3d4e5f6a7b8c9d0",
                        "title": "Amazing Tea Discovery",
                        "content": "I discovered this amazing tea blend with notes of jasmine and honey.",
                        "authorId": "64f5a1b2c3d4e5f6a7b8c9d1",
                        "categoryId": {
                          "_id": "64f5a1b2c3d4e5f6a7b8c9d2",
                          "name": "Green Tea"
                        },
                        "likes": [],
                        "likesCount": 0,
                        "score": 0,
                        "createdAt": "2025-09-29T12:00:00.000Z",
                        "updatedAt": "2025-09-29T12:00:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Category not found"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get posts with pagination and filtering",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of posts per page"
          },
          {
            "in": "query",
            "name": "categoryId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by category ID"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "relevance",
                "freshness",
                "likeCount"
              ],
              "default": "createdAt"
            },
            "description": "Sort field"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "description": "Sort order"
          },
          {
            "in": "query",
            "name": "algorithm",
            "schema": {
              "type": "string",
              "enum": [
                "base",
                "trend"
              ],
              "default": "base"
            },
            "description": "Scoring algorithm for relevance-based sorting (A/B: base vs trend)"
          }
        ],
        "responses": {
          "200": {
            "description": "Posts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "$ref": "#/components/schemas/PaginationResponse"
                    }
                  ]
                },
                "examples": {
                  "paginatedPosts": {
                    "summary": "Paginated posts response example",
                    "value": {
                      "success": true,
                      "data": {
                        "posts": [
                          {
                            "_id": "64f5a1b2c3d4e5f6a7b8c9d0",
                            "title": "Amazing Tea Discovery",
                            "content": "I discovered this amazing tea blend...",
                            "likesCount": 15,
                            "score": 9.2
                          },
                          {
                            "_id": "64f5a1b2c3d4e5f6a7b8c9d3",
                            "title": "Brewing Guide",
                            "content": "How to brew the perfect cup",
                            "likesCount": 5,
                            "score": 6.1
                          }
                        ],
                        "pagination": {
                          "page": 1,
                          "limit": 20,
                          "offset": 0,
                          "total": 100,
                          "totalPages": 5,
                          "hasNext": true,
                          "hasPrevious": false
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/posts/bulk": {
      "post": {
        "summary": "Create multiple posts in bulk",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCreatePostsRequest"
              },
              "examples": {
                "sampleBulk": {
                  "summary": "Bulk create example",
                  "value": {
                    "posts": [
                      {
                        "title": "Post 1",
                        "content": "Content 1",
                        "categoryId": "64f5a1b2c3d4e5f6a7b8c9d2"
                      },
                      {
                        "title": "Post 2",
                        "content": "Content 2",
                        "categoryId": "64f5a1b2c3d4e5f6a7b8c9d2"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Posts created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "posts": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Post"
                              }
                            },
                            "summary": {
                              "type": "object",
                              "properties": {
                                "total": {
                                  "type": "number"
                                },
                                "successful": {
                                  "type": "number"
                                },
                                "failed": {
                                  "type": "number"
                                },
                                "errors": {
                                  "type": "array",
                                  "items": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                },
                "examples": {
                  "bulkCreated": {
                    "summary": "Bulk create response",
                    "value": {
                      "success": true,
                      "data": {
                        "posts": [
                          {
                            "_id": "64f5a1b2c3d4e5f6a7b8c9d0",
                            "title": "Post 1",
                            "content": "Content 1",
                            "likesCount": 0
                          },
                          {
                            "_id": "64f5a1b2c3d4e5f6a7b8c9d1",
                            "title": "Post 2",
                            "content": "Content 2",
                            "likesCount": 0
                          }
                        ],
                        "summary": {
                          "total": 2,
                          "successful": 2,
                          "failed": 0,
                          "errors": []
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "207": {
            "description": "Some posts created, some failed (Multi-Status)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, too many posts, or all posts failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/posts/{id}": {
      "get": {
        "summary": "Get a single post by ID",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Post ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Post retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Post"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Post not found"
          }
        }
      }
    },
...
  },
...
}